Refactor import path for XMTP client initialization
