+++ bots/key-check/index.ts
@@ -7,7 +7,7 @@ import {
   type DecodedMessage,
   type GroupMember,
 } from "version-management/client-versions";
-import { initializeClient } from "../helpers/xmtp-handler";
+import { initializeClient } from "../helpers/xmtp-skills";
 
 // Get XMTP SDK version from package.json
 const require = createRequire(import.meta.url);
